C2C-223: Changed type of cause_of_death column to INT because of issues with H6 data. (#3)

+++ analytics/dsl/export/tables/patients.sql
@@ -32,7 +32,7 @@ CREATE TABLE patients (
     attributes VARCHAR,
     dead BOOLEAN,
     death_date TIMESTAMP,
-    cause_of_death BIGINT,
+    cause_of_death INT,
     creator BIGINT,
     date_created TIMESTAMP,
     person_voided BOOLEAN,
